Handshake tests now validate lengths, and has a benchmark.
The benchmark (and the tests for that matter) can fail due to false
positives in the bloom filter.

The combined time for Alice + Bob's portion of the key exchange on an
i7-5600U is ~1.1 ms (approx 550 usec/side), which is a bit slower than
I'd like but still acceptable, given the massively over engineered
handshake.