Optionally force client side PLPMTUD.
The client side can now force-enable RFC 4821 Packetization-Layer PMTUD
on certain platforms (Linux).  This is off by default, and should only
be enabled in certain corner cases.

For an example of firewall braindamage this attempts to solve see:
https://labs.riseup.net/code/issues/9268

The bug is inconclusive as to if this will actually help, but it's easy
to allow, and the only other solution is "complain to the people that
mis-configured the firewall".

Fixes #7