Add `Obfs4PacketIAT` (equivalent to `iat-mode=2`).
I'm still uncertain how effective this is, so servers will not offer it
by default, though clients will prefer it over either of the burst
modes if it is available.

The code for it basically already existed in the form of the short
write obfuscation handler, so offering it is easy enough to do.